+++ stattrack/commands.py
@@ -1,11 +1,15 @@
 import datetime
+import json
+import logging
 from io import StringIO
+from time import monotonic
 from typing import Iterable, Optional, Union
 
 import discord
+import pandas as pd
 from discord.ext.commands.cooldowns import BucketType
 from redbot.core import commands
-from redbot.core.utils.chat_formatting import box, humanize_timedelta
+from redbot.core.utils.chat_formatting import box, humanize_number, humanize_timedelta
 
 from stattrack.abc import MixinMeta
 from stattrack.converters import (
@@ -14,6 +18,8 @@ from stattrack.converters import (
     TimespanConverter,
     UserGraphConverter,
 )
+
+log = logging.getLogger("red.vex.stattrack")
 
 DEFAULT_DELTA = datetime.timedelta(days=1)
 
@@ -29,16 +35,32 @@ class StatTrackCommands(MixinMeta):
         *,
         more_options: bool = False,
     ):
+        comstart = monotonic()
         if self.df_cache is None:
             return await ctx.send("This command isn't ready yet. Try again in a few seconds.")
-        await ctx.trigger_typing()  # wont be that long
         if ylabel is None:
             ylabel = title
         df = self.df_cache[label]
         if len(df) < 2:
             return await ctx.send("I need a little longer to collect data. Try again in a minute.")
 
-        graph = await self.plot(df, delta, ylabel)
+        # index data to desired delta
+        now = datetime.datetime.utcnow().replace(microsecond=0, second=0)
+        start = now - delta
+        start = max(start, df.first_valid_index())
+
+        delta_max_points = (now - start).total_seconds() / 60
+
+        maxpoints = await self.config.maxpoints()
+        if delta_max_points > 1440 and maxpoints != -1:  # 1 day
+            frequency = int(delta_max_points // maxpoints)
+        else:
+            frequency = 1
+        expected_index = pd.date_range(start=start, end=now, freq=f"{frequency}min")
+        df = df.reindex(index=expected_index)
+
+        async with ctx.typing():
+            graph = await self.plot(df, ylabel)
 
         if delta == datetime.timedelta(days=9000):  # "all" was entered and was replaced with 9k
             str_delta = " all time"
@@ -57,16 +79,61 @@ class StatTrackCommands(MixinMeta):
             )
         embed.set_footer(text="Times are in UTC")
         embed.set_image(url="attachment://plot.png")
-        await ctx.send(file=graph, embed=embed)
+
+        msg = await ctx.send(file=graph, embed=embed)
+
+        end = monotonic()
+        debug_info = {
+            "plot_msg": msg.id,  # message id of sent plot
+            "alldatapoints": len(self.df_cache),  # datapoints in the all-time dataframe
+            "maxpoints": maxpoints,  # user set max points to plot on a graph
+            "mins": delta_max_points,  # the amount of minutes in the delta
+            "points_plotted": len(df),  # valid datapoints in the delta dataframe
+            "wanted_frequency": frequency,  # wanted frequency of the plot
+            "plotted": label,  # metrics plotted
+            "comtime": end - comstart,  # time taken to plot the graph
+        }
+
+        log.debug(f"Plot finished, info: {debug_info}")
+        self.last_plot_debug = debug_info
 
     @commands.cooldown(10, 60.0, BucketType.user)
     @commands.group()
     async def stattrack(self, ctx: commands.Context):
         """View my stats."""
 
+    @commands.is_owner()
     @stattrack.command(hidden=True)
     async def raw(self, ctx, var):
         await ctx.send(box(str(self.df_cache[var])))
+
+    @commands.is_owner()
+    @stattrack.command(hidden=True)
+    async def devimport(self, ctx: commands.Context):
+        """
+        Import data from a JSON string, orient "split".
+
+        This is for development purposes only.
+
+        Please attack a `.json` file.
+        """
+        self.df_cache = self.df_cache.append(
+            pd.read_json(await ctx.message.attachments[0].read(), orient="split")
+        )
+        await self.driver.write(self.df_cache)
+        await ctx.send("Done.")
+
+    @commands.is_owner()
+    @stattrack.command(hidden=True)
+    async def debug(self, ctx: commands.Context):
+        """
+        View debug info for the last plot.
+
+        This is for development purposes only.
+        """
+        if self.last_plot_debug is None:
+            await ctx.send("No plot has been made yet.")
+        await ctx.send(box(json.dumps(self.last_plot_debug, indent=4), "json"))
 
     @commands.is_owner()
     @stattrack.group()
@@ -116,6 +183,35 @@ class StatTrackCommands(MixinMeta):
             return
         fp.seek(0)
         await ctx.send("Here is your file.", file=discord.File(fp, "stattrack.csv"))  # type:ignore
+
+    @commands.is_owner()
+    @stattrack.command()
+    async def maxpoints(self, ctx: commands.Context, maxpoints: int):
+        """
+        Set the maximum number of points to plot. This affects the speed of graph plotting.
+
+        The default value is 25k (25000).
+
+        The more points you plot, the slower the plotting time will be.
+
+        This setting only affects graphs that are a longer timespan (1 month+).
+
+        Set maxpoints to -1 to disable this feature, therefore always plotting all points.
+
+        Otherwise, maxpoints must be at least 1k (1440).
+
+        **Examples:**
+            - `[p]stattrack maxpoints 10000` - plot up to 10k points
+            - `[p]stattrack maxpoints 75000` - plot up to 75k points
+            - `[p]stattrack maxpoints 1440` - the minimum value possible
+            - `[p]stattrack maxpoints 25000` - the default value
+            - `[p]stattrack maxpoints -1` - disable, always plot all points
+        """
+        if maxpoints < 1440 and maxpoints != -1:
+            await ctx.send("The minimum value is 1440.")
+            return
+        await self.config.maxpoints.set(maxpoints)
+        await ctx.send(f"Done, the maximum points to plot is now {humanize_number(maxpoints)}.")
 
     @stattrack.command(aliases=["ping"])
     async def latency(self, ctx: commands.Context, timespan: TimespanConverter = DEFAULT_DELTA):